install and finish Activity - NavigationContainer and Activity - createNativeStackNavigator
